+++ src/main/java/br/com/duosdevelop/vb/igrejaalocacao/repositories/CelulaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CelulaRepository extends JpaRepository<Celula, Integer> {
+public interface CelulaRepository extends JpaRepository<Celula, Long> {
 }
