+++ src/main/java/br/com/duosdevelop/vb/igrejaalocacao/repositories/CultoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CultoRepository extends JpaRepository<Culto, Integer> {
+public interface CultoRepository extends JpaRepository<Culto, Long> {
 }
