+++ src/main/java/br/com/duosdevelop/vb/igrejaalocacao/repositories/DiscipuladorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DiscipuladorRepository extends JpaRepository<Discipulador, Integer> {
+public interface DiscipuladorRepository extends JpaRepository<Discipulador, Long> {
 }
