+++ src/main/java/br/com/duosdevelop/vb/igrejaalocacao/repositories/RedeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RedeRepository extends JpaRepository<Rede, Integer> {
+public interface RedeRepository extends JpaRepository<Rede, Long> {
 }
